Make testSuite compile on old Scala versions
