+++ test-suite/js/src/test/scala-2.13/org/scalajs/testsuite/library/WrappedDictionaryToTest.scala
@@ -1,0 +1,23 @@
+/*                     __                                               *\
+**     ________ ___   / /  ___      __ ____  Scala.js Test Suite        **
+**    / __/ __// _ | / /  / _ | __ / // __/  (c) 2013, LAMP/EPFL        **
+**  __\ \/ /__/ __ |/ /__/ __ |/_// /_\ \    http://scala-js.org/       **
+** /____/\___/_/ |_/____/_/ | |__/ /____/                               **
+**                          |/____/                                     **
+\*                                                                      */
+package org.scalajs.testsuite.library
+
+import scala.scalajs.js
+
+import org.junit.Assert._
+import org.junit.Test
+
+class WrappedDictionaryToTest {
+
+  @Test def to_T(): Unit = {
+    val dict = js.Dictionary("a" -> "a", "b" -> 6, "e" -> js.undefined)
+    val list = dict.to(List)
+    assertEquals(3, list.size)
+  }
+
+}
