+++ test-suite/js/src/test/scala/org/scalajs/testsuite/library/ArrayOpsTest.scala
@@ -113,9 +113,4 @@ class ArrayOpsTest {
     assertArrayEquals(array.toArray, list.toArray)
   }
 
-  @Test def to_T_issue_843(): Unit = {
-    val array = js.Array(1,2,1,3,1,10,9)
-    val list = array.to(List)
-    assertArrayEquals(array.toArray, list.toArray)
-  }
 }
