+++ test-suite/js/src/test/scala/org/scalajs/testsuite/library/WrappedArrayTest.scala
@@ -107,10 +107,4 @@ class WrappedArrayTest {
     assertEquals(List(1,2,1,3,1,10,9), list)
   }
 
-  @Test def to[T](): Unit = {
-    val seq: collection.Seq[Int] = js.Array(1,2,1,3,1,10,9)
-    val list = seq.to(List)
-    assertEquals(List(1,2,1,3,1,10,9), list)
-  }
-
 }
