+++ test-suite/js/src/test/scala/org/scalajs/testsuite/library/WrappedDictionaryTest.scala
@@ -95,10 +95,4 @@ class WrappedDictionaryTest {
     assertEquals(3, list.size)
   }
 
-  @Test def to_T(): Unit = {
-    val dict = js.Dictionary("a" -> "a", "b" -> 6, "e" -> js.undefined)
-    val list = dict.to(List)
-    assertEquals(3, list.size)
-  }
-
 }
